requirements.txt
Encoder==1.1
pigpio==1.78
pigpio-encoder==0.2.4
RPi.GPIO==0.7.1
tk==0.1.0
